==> Script/01.Github_Download.R <==

# 01_깃허브 연습폴더 다운로드
download.file(url = "https://github.com/cmman75/Open_data_R_with_Shiny/archive/master.zip" , destfile = "master.zip")

# 02_다운로드 파일 압축풀기
unzip(zipfile = "master.zip")

# 03_Open_data_R_with_Shiny_main 로 들어가기
